requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.1
Django==4.0
django-environ==0.8.1
django-simple-history==3.0.0
djangorestframework==3.13.0
pycparser==2.21
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
